Provide format-agnostic wrapper of output writing
OutputWriter is an interface for writing output in either Eclipse- or
VTK-compatible format, and contains a routine for getting a writer of
that interface based on the parameters passed to the simulator.

This enables us to toggle various formats on and off for the simulators
through parameters, while the source code only use the same interface.