Enable automatic synchronization of state
The default is now to synchronize state (between the 2D and 3D view) for
every timestep. If we know that we are calling it through the wrapper,
then this is easy to disable. If we are passing the type of the wrapper
as a drop-in replacement for the real simulator (for performance or
stability reasons), then we probably want to have the state automatically
synchronized because it is hard to fudge in code that does it otherwise.